edit map and control service to put logic about legend control in service
